Conflicts:
	ddl/postgres/deapp/de_subject_protein_data.sql

+++ ddl/postgres/tm_wz/wt_subject_microarray_med.sql
@@ -3,8 +3,8 @@
 --
 CREATE TABLE wt_subject_microarray_med (
     probeset_id bigint,
-    intensity_value bigint,
-    log_intensity bigint,
+    intensity_value numeric,
+    log_intensity numeric,
     assay_id bigint,
     patient_id bigint,
     sample_id bigint,
@@ -13,9 +13,9 @@ CREATE TABLE wt_subject_microarray_med (
     timepoint character varying(100),
     pvalue double precision,
     num_calls bigint,
-    mean_intensity bigint,
-    stddev_intensity bigint,
-    median_intensity bigint,
-    zscore bigint
+    mean_intensity numeric,
+    stddev_intensity numeric,
+    median_intensity numeric,
+    zscore double precision
 );
