+++ ddl/postgres/tm_wz/wt_subject_mrna_probeset.sql
@@ -5,7 +5,7 @@ CREATE TABLE wt_subject_mrna_probeset (
     probeset_id bigint,
     expr_id character varying(500),
     intensity_value numeric,
-    num_calls numeric,
+    num_calls bigint,
     pvalue numeric,
     assay_id bigint,
     patient_id bigint,
